added inventory page with previous functions (add/remove/update bead inventory) | added empty create/index as a starting point
