+++ views/create/index.jade
@@ -4,90 +4,6 @@ extends ../layout
 block content
 
   a(href="/") Homepage
-  h1 Bead Inventory:
-  table
-    tr
-      th(class="small")
-        button(type="submit" onclick="updateCheckboxes()")
-      th id 
-      th desc
-      th qty
-      
-    each bead in beads
-      tr
-        td(class="small")
-          input(type="checkbox" id="#{bead.id}_checkbox")
-        td #{bead.id}
-        td #{bead.desc}
-        td 
-          form(action="", method="post")
-            input(type="hidden", name="act", value="update")
-            input(type="hidden", name="id", value="#{bead.id}")
-            input(type="text" name="qtyText" value="#{bead.qty}")
-        
-        //- - if(admin == 1) {
-        //-   td #{bead.msg}
-        //- - }
-        
-  br
-  h6(style="color: green")=success
-  br
-  h6 Add bead to inventory
-  h6(style="color: red")=error
+  h1 Create Necklace:
   
-  p=checkList
-  form(action="", method="post")
-    input(type="hidden", name="act", value="add")
-    input(type="text", name="id", placeholder="ID")
-    input(type="text", name="desc", placeholder="Description")
-    input(type="text", name="qty", placeholder="Quantity")
-    
-    br
-    input(type="submit")
-  
-  script.
-    var checkList;
-    function updateCheckboxes() {
-      var form = document.createElement("form");
-      form.setAttribute("method", "post");
-      form.setAttribute("action", "");
-      
-      // get status and id for items to be removed
-      checkList = document.querySelectorAll('[id*="_checkbox"]');
-      console.log(checkList);
-      var removeList = [];
-      
-      // get anything that needs to be removed
-      for (i = 0; i < checkList.length; i++){
-        if (checkList[i].checked) {
-          // remove the _checkbox from the id
-          var cleanStr = checkList[i].id.substring(0, checkList[i].id.indexOf('_'));
-          removeList.push(cleanStr);
-          //- removeList[cleanStr] = checkList[i].checked;
-          //- console.log(`${cleanStr} ${checkList[i].checked}`);
-        }
-      }
-      console.log(removeList);
-      // stringify for form submission
-      //- var stringForm = JSON.stringify(removeList);
-      console.log('====');
-      console.log(removeList);
-      var hiddenField = document.createElement("input");
-      hiddenField.setAttribute("type", "hidden");
-      hiddenField.setAttribute("name", "removeList");
-      //- hiddenField.setAttribute("value", stringForm);
-      hiddenField.setAttribute("value", removeList);
-      form.appendChild(hiddenField);
-      
-      hiddenField = document.createElement("input");
-      hiddenField.setAttribute("type", "hidden");
-      hiddenField.setAttribute("name", "act");
-      hiddenField.setAttribute("value", "remove");
-      form.appendChild(hiddenField);
-      
-      // adding field->form->body then submit
-      document.body.appendChild(form);
-      form.submit();
-    }
-
