+++ views/index.jade
@@ -3,5 +3,8 @@ extends layout
 block content
   h1 #{title}
   //- button(class="btn btn-primary") btn
-  p #{msg}
+  p To create a necklace
     a(href="/create") click here
+  
+  p Update inventory
+    a(href="/inventory") here
