+++ views/inventory/success.jade
